wgengine/router: split out from wgengine.
The router implementations are logically separate, with their own API.
