+++ wgengine/router/ifconfig_windows.go
@@ -3,7 +3,7 @@
  * Copyright (C) 2019 WireGuard LLC. All Rights Reserved.
  */
 
-package wgengine
+package router
 
 import (
 	"bytes"
