+++ wgengine/router/router_default.go
@@ -4,15 +4,13 @@
 
 // +build !windows,!linux,!darwin,!openbsd,!freebsd
 
-package wgengine
+package router
 
 import (
 	"github.com/tailscale/wireguard-go/device"
 	"github.com/tailscale/wireguard-go/tun"
 	"tailscale.com/types/logger"
 )
-
-const DefaultTunName = "tailscale0"
 
 func newUserspaceRouter(logf logger.Logf, tunname string, dev *device.Device, tuntap tun.Device, netChanged func()) Router {
 	return NewFakeRouter(logf, tunname, dev, tuntap, netChanged)
